feat: Add sniffer kernel module to the modules directory.
This commit adds the sniffer module to the modules directory of the
kernel. This is just a referecen, as this code has to be compile along
side the kernel ir order for it to be part of it.

To see how this is done, refer to the compiline the kernel session in
the README file.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>