feat: Implement sniffer lib and client
This commit implements the sniffer library that can read the files
created by the sniffer kernel module.

This commit also implements the sniffer client. This client has the
sniffer library as its dependecy and uses it to issue commands to the
library expecting the library's reponse.

At this momment, the library only returns the RSSI value.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>