feat: Refactor code to trap the call to get_network_info
This commit changes the place where the command is being parsed. Now the
while that goes though the command traps the execution of the callback
function and prevents any informatino mismaatch as well as any
nullpointer exceptions that may be caused from accessing the structs
pointers at the wrong time.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>