feat: Implement sniffer HAL
This commit implements the Sniffer HAL. This is the interface between
the upper layer platform and the bottom layer. It will be responsible
for serving as a bridge among the service client and the sniffer
library.

This commit also implements the SELinux rules the the Sniffer HAL needs
to function.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>