chore: Change project structure
This commit change the project structure for the framework related
classes. Now we have a frameworks/ directory that wraps everything
related to framework.

There's also an Android.bp which takes care of the most important
libraries in the framework. This does not block the creation of
directories with specific Android.bp files.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>