fix: Fix #4 board sometimes not responding
RCA: Sometimes the captured packet at the moment of the board request is
not of type WIFI_PKT_MGMT and subtype BEACON. For these cases there was
no response whatsoever, causing the kernel module to wait until a
timeout occurs.

SOL: Now, the board writes to the Serial a default string naming the
network as UNKNOWN (SSID) and giving a 0 as its RSSI value.

This commit also brings better documentation and changes the names
of some structs, as well as improviments regarding the code identation.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>