refactor: #8 - Refactor to add WIFI_PKT_MGMT filter
This commit refactors the esp32 code to make use of the esp_wifi_set_promiscuous_filter
API. This API can filter out the a set of packets given a mask. We are
using to filter the WIFI_PKT_MGMT, avoiding calling the callback
function for those packets that are not of this type.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>