feat: #9 - Implement HAL 2.0
This commit implements the HAL Service 2.0.

This new HAL now defines a NetworkInfo interface that will be used by
both lower and upper layer to describe the information the board
captures.

The new SnifferNetworkInfo class now must be used by the applications to
be able to access the network information easily.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>