fix: #10 - Fix SnifferNetworkInfo's equals method
RCA: The SnifferNetworkInfo's equals method was using both, SSID and
RSSI to compare if two SnifferNetworkInfo objects were equal. This is
not ideal as the RSSI value can varie from time to time to the same
network.

SOL: Use only the SSID to decide if two SnifferNetworkInfo object are
equal.

Signed-off-by: Malbolge <victor.gomes@icomp.ufam.edu.br>